Also healing on the <share://> subjects
Also heal the physical files, so that the files-consumer does not have
to do that. This was previously broken or intentionally done in a
certain way that would ignore the triples from the file-consumer graphs.
I think healing on physical files is needed for the edge case
dispatching rules change and a submission is also added to another
bestuurseenheid.